페이퍼 타입 분리
