+++ Dayol-iOS/UI/DiaryPaperModal/PaperAdd/ViewModel/AddPaperContentViewModel.swift
@@ -30,17 +30,19 @@ class AddPaperContentViewModel {
         
         // TODO: 모델 init 간편화 필요
         // TODO: Model을 따로두고 Model이 Entity와 소통하도록 변경해야함
+        let paperSize = PaperOrientationConstant.size(orentantion: model.orientation)
         let paperModel: Paper = Paper(id: DYTestData.shared.currentPaperId,
                                       diaryId: diaryId,
                                       title: model.title,
                                       pageCount: 1,
                                       orientation: model.orientation,
-                                      type: .init(value: model.paperType),
-                                      size: PaperOrientationConstant.size(orentantion: model.orientation),
+                                      type: model.paperType,
+                                      width: Float(paperSize.width),
+                                      height: Float(paperSize.height),
                                       thumbnail: nil,
                                       drawCanvas: Data(),
                                       contents: [],
-                                      date: model.paperType.date)
+                                      date: Date.now)
         
         DYTestData.shared.addPaper(paperModel)
     }
