+++ Dayol-iOS/UI/DiaryPaperModal/PaperList/ViewModel/PaperListContentViewModel.swift
@@ -26,7 +26,7 @@ class PaperListContentViewModel {
 
     var cellModels: [CellModel] {
         return DYTestData.shared.paperList.map {
-            let paperType = PaperType(rawValue: $0.type, date: $0.date) ?? .muji
+            let paperType = $0.type
             let orientaion = Paper.PaperOrientation(rawValue: $0.orientation) ?? .portrait
             return CellModel(id: $0.id, isStarred: false, orientation: orientaion, paperType: paperType, thumbnailData: $0.thumbnail)
         }
